This should presumably fix gcc 7.5 build
