+++ sycl/source/detail/accessor_impl.hpp
@@ -10,9 +10,9 @@
 
 #include <sycl/access/access.hpp>
 #include <sycl/accessor.hpp>
-#include <sycl/accessor_image.hpp>
 #include <sycl/detail/export.hpp>
 #include <sycl/id.hpp>
+#include <sycl/image.hpp>
 #include <sycl/property_list.hpp>
 #include <sycl/range.hpp>
